+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ###Version: 0.11.7
 - ```new``` Landscape layout for calculator. (**@i906**)
 - ```fix``` Can edit main currency.
+- ```fix``` Better overview landscape support.
 
 ###Version: 0.11.6
 - ```new``` Tags are ordered alphabetically.